+++ __init__.py
@@ -21,7 +21,7 @@ mouse_ = shared.mouse_
 
 print('Current Expy: ', os.path.dirname(os.path.abspath(__file__)))
 
-def start(setting_file='setting.txt', fullscreen=True, winsize=(800, 600), mouse_visible=False, normal_font_size=20, stim_font_size=None, distance=60, diag=23, angel=2.5, font_color=C_white, background_color=C_gray, sample_rate=44100, port='', vsync=False):
+def start(setting_file='setting.txt', fullscreen=True, winsize=(800, 600), mouse_visible=False, normal_font_size=20, stim_font_size=None, distance=60, diag=15, angel=1.6, font_color=C_white, background_color=C_gray, sample_rate=44100, port='', vsync=True):
     '''
     Initialize the experiment.
     Note: todo.
@@ -55,7 +55,7 @@ def start(setting_file='setting.txt', fullscreen=True, winsize=(800, 600), mouse
     port: str, or hex number (default:'') 
         Port name used to send trigger.
         Use str on serial port, and hex on parallel port 
-    vsync: True, or False (default)
+    vsync: True (default), or False 
         Vertical retrace synchronisation
 
     Returns