+++ stim/display.py
@@ -23,16 +23,19 @@ def clear(debugging=True):
         shared.win.clear()
         shared.win.flip() 
 
-def show(out_time=False, clean_screen=True, backup=None, debugging=True):
+
+def show(out_time=False, clean_screen=True, stop_signal=None, backup=None, debugging=True):
     '''
     Display current canvas buffer, and keep the display during a limited period.
 
     Parameters
     ----------
     out_time: num(>0), False(default)
-        The time limit of current function. (unit: second) 
+        The time limit of current function. (unit: second)
     clean_screen: True(default), False
         Whether clear the screen after get the screen or not. 
+    stop_signal: None (default), str
+        If the stop_signal is str, this method would be blocked until a stop_signal was received from the 0.0.0.0:36
     backup: None, or a screen backup
         Give a prepared screen to display
 
@@ -48,11 +51,20 @@ def show(out_time=False, clean_screen=True, backup=None, debugging=True):
         shared.win.flip()
         shared.need_update = False
 
+        # if not shared.start_tp:
+        #     shared.start_tp = shared.time.time()
+
     # shared.pg.display.flip()
     if out_time:
         waitForResponse(shared.key_.ENTER, out_time)
-        if clean_screen:
-            clear(debugging)
+    if stop_signal!=None:
+        s = stop_signal.encode(encoding='utf_8', errors='strict')
+        while True:
+            if shared.net_port_state == s:
+                # print(shared.net_port_state)
+                break
+    if clean_screen:
+        clear(debugging)
 
 def getScreen(clean_screen=True):
     '''