+++ test/read_setting.py
@@ -1,7 +1,7 @@
 # coding:utf-8
 ##### package test #####
 import sys
-sys.path = ['../']+sys.path
+sys.path = ['../../']+sys.path
 ################
 
 from expy import *  # Import the needed functions
@@ -15,7 +15,5 @@ print(shared.setting['timing_set'])  # Print the setting of 'timing_set' part
 
 'Calling "timing" function to get a certain time value'
 print(timing('ITI'))  # Print 'ITI' value (we specify 500 for it in "setting.txt")
-# Print 'fix' value (we specify 800~1400 range for it in "setting.txt")
-print(timing('fix'))
 # Print 'fix' value (we specify 800~1400 range for it in "setting.txt")
 print(timing('fix'))