+++ test/send_trigger.py
@@ -5,25 +5,38 @@ sys.path = ['../']+sys.path
 ################
 print(sys.version)
 
-from expy import *  # Import the needed functions
+# from expy import *  # Import the needed functions
 
 'Parallel Port'
-# start(port=0xC100)
-# sendTrigger(0, mode='P')
+start(port=0xC100)
+sendTrigger(0, mode='P')
 
 'Serial Port'
 start(port='COM5')
 sendTrigger('m$', mode='S')
 
 # import serial
+# import time
 # ser = serial.Serial(baudrate=115200)
-# ser.port = 'COM1' # set the port
+# ser.port = 'COM5' # set the port
 # try:
 #     ser.open()
 # except:
 #     print('Could not open port')
 
-# ser.write(b'something') # send a string directly
+# time.sleep(5)
+# # ser.write(chr(64))  # send a string directly
+# print('1')
+# for i in range(13,25):
+#     print(i)
+#     ser.write(b'\x4d'+bytes([i]))
+#     time.sleep(1)
+
+
+# print(ser.read())
+# ser.write()
+# ser.write(chr(64))
+
 
 # tg = 'something'
 # ser.write(bytes(tg,encoding='utf8')) # send a string which might change